requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Core dependencies
setuptools==69.5.1

# Image processing - compatible versions
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.6.0.66

# OCR dependencies
paddlepaddle==2.6.1
paddleocr==2.7.3
shapely==2.0.4
pyclipper==1.3.0.post5

# HTTP requests
requests==2.31.0